Switch order of anchor_text and footnote_content
I want to be able to include quotes and html in my footnote content,
but was getting errors when it was included as an argument to the
liquid shortcode.

However, formatted/templated text is allowed in the content between
paired shortcodes, so it makes more sense for me to put the footnote
content there, and assume that the anchor will be a single word or phrase.